added sample data and card changes
