+++ src/components/header/Header.tsx
@@ -5,7 +5,7 @@ import { FiHome, FiUser, FiHeart, FiShoppingBag } from "react-icons/fi";
 
 export const Header: FC = () => {
   return (
-    <nav className="sticky top-0 flex p-2 w-full justify-center md:justify-between">
+    <nav className="sticky z-50 top-0 flex p-5 bg-white w-full justify-center md:justify-between">
       <Image src="/recomm.svg" width={100} alt="logo" height={50} />
       <ul className="links hidden md:flex gap-10">
         <Link href="/" className="link-item flex gap-1">
