+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export * from "./header/Header";
 export * from "./footer/Footer";
 export * from "./bottom-navigation/BottomNavigation";
-export * from "./product-list/ProductList";
+export * from "./product-list/ProductCard";
 export * from "./ad-section/ADSection";
